% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_abstract_data_wrapper.R
\name{abstract_data_wrapper}
\alias{abstract_data_wrapper}
\title{An abstract data wrapper}
\usage{
abstract_data_wrapper(id, cell_ids, cell_info = NULL, ...)
}
\arguments{
\item{id}{A unique identifier for the data}

\item{cell_ids}{The ids of the cells.}

\item{cell_info}{Optional meta-information pertaining the cells.}

\item{...}{Extra information to be stored in the wrapper.}
}
\description{
An abstract data wrapper
}
