% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_class.R
\name{add_class}
\alias{add_class}
\title{Add class to object whilst keeping the old classes}
\usage{
add_class(x, class)
}
\arguments{
\item{x}{a \R object}

\item{class}{A character vector naming classes}
}
\description{
Add class to object whilst keeping the old classes
}
\examples{
library(purrr)
l <- list(important_number = 42) \%>\% add_class("my_list")
}
