% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_dimred_to_wrapper.R
\name{add_dimred_to_wrapper}
\alias{add_dimred_to_wrapper}
\title{Add a dimensionality reductio n to a data wrapper}
\usage{
add_dimred_to_wrapper(data_wrapper, dimred, dimred_milestone = NULL, ...)
}
\arguments{
\item{data_wrapper}{A data wrapper to extend upon.}

\item{dimred}{The dimensionality reduction matrix.}

\item{dimred_milestone}{An optional dimensionality reduction of the milestones.}

\item{...}{extra information to be stored in the wrapper}
}
\description{
Add a dimensionality reductio n to a data wrapper
}
