% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{all_in}
\alias{all_in}
\alias{\%allin\%}
\title{Check whether a vector are all elements of another vector}
\usage{
all_in(x, table)

x \%allin\% table
}
\arguments{
\item{x}{The values to be matched.}

\item{table}{The values to be matched against.}
}
\description{
Check whether a vector are all elements of another vector
}
