% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance.R
\name{calculate_distance}
\alias{calculate_distance}
\alias{list_distance_metrics}
\alias{calculate_similarity}
\alias{list_similarity_metrics}
\title{Calculate (column-wise) distances/similarity between two matrices}
\usage{
calculate_distance(x, y = NULL, metric = c("cosine", "pearson",
  "spearman", "euclidean", "manhattan"), margin = 1)

list_distance_metrics()

calculate_similarity(x, y = NULL, margin = 1, metric = c("cosine",
  "pearson", "spearman"))

list_similarity_metrics()
}
\arguments{
\item{x}{A numeric matrix, dense or sparse.}

\item{y}{(Optional) a numeric matrix, dense or sparse, with \code{nrow(x) == nrow(y)}.}

\item{metric}{Which distance metric to use. Options are: \code{"cosine"}, \code{"pearson"}, \code{"spearman"}, \code{"euclidean"}, and \code{"manhattan"}.}

\item{margin}{Which margin to use for the pairwise comparison. 1 => rowwise, 2 => columnwise.}
}
\description{
These matrices can be dense or sparse.
}
\examples{
## Generate two matrices with 50 and 100 samples
library(Matrix)
x <- Matrix::rsparsematrix(50, 1000, .01)
y <- Matrix::rsparsematrix(100, 1000, .01)

dist_euclidean <- calculate_distance(x, y, metric = "euclidean")
dist_manhattan <- calculate_distance(x, y, metric = "manhattan")
dist_spearman <- calculate_distance(x, y, metric = "spearman")
dist_pearson <- calculate_distance(x, y, metric = "pearson")
dist_angular <- calculate_distance(x, y, metric = "cosine")
}
