% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{check_packages}
\alias{check_packages}
\title{Check which packages are installed}
\usage{
check_packages(dependencies)
}
\arguments{
\item{dependencies}{The names of the packages to be installed}
}
\description{
Check which packages are installed
}
