% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_like_distance.R
\name{compute_emlike_dist}
\alias{compute_emlike_dist}
\title{Calculate Earth Mover's Distance between cells in a trajectory}
\usage{
compute_emlike_dist(traj)
}
\arguments{
\item{traj}{the trajectory}
}
\description{
Calculate Earth Mover's Distance between cells in a trajectory
}
