% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_tented_geodesic_distances.R
\name{compute_tented_geodesic_distances}
\alias{compute_tented_geodesic_distances}
\title{Calculate geodesic distances between cells in a trajectory, taking into account tents}
\usage{
compute_tented_geodesic_distances(trajectory)
}
\arguments{
\item{trajectory}{the trajectory}
}
\description{
Calculate geodesic distances between cells in a trajectory, taking into account tents
}
