% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{dangerous_assignin_namespace}
\alias{dangerous_assignin_namespace}
\title{Assigning an object in a different namespace.}
\usage{
dangerous_assignin_namespace(x, value, ns, pos = -1,
  envir = as.environment(pos))
}
\description{
It was copied from \code{\link[utils]{assignInNamespace}},
but a certain check was removed so that the set.seed function can be overwritten.
}
\details{
This function is not to be used aside from overriding the set.seed function within the
TI evaluation.
}
