% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance.R
\name{euclidean_distance}
\alias{euclidean_distance}
\alias{correlation_distance}
\title{These functions will be removed soon}
\usage{
euclidean_distance(x, y = NULL)

correlation_distance(x, y = NULL)
}
\arguments{
\item{x}{A numeric matrix, dense or sparse.}

\item{y}{(Optional) a numeric matrix, dense or sparse, with \code{nrow(x) == nrow(y)}.}
}
\description{
Use \code{\link[=calculate_distance]{calculate_distance()}} instead.
}
