% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{dynutils}
\alias{dynutils}
\alias{dynutils-package}
\title{Common functionality for the dynverse packages}
\description{
Provides common functionality for the dynverse packages.
dynverse is created to support the development, execution, and benchmarking of trajectory inference methods.
For more information, check out \url{https://github.com/dynverse/dynverse}.
}
\section{Manipulation of lists}{

\itemize{
  \item{\code{\link{add_class}}: Add a class to an object}
  \item{\code{\link{extend_with}}: Extend list with more data}
}
}

\section{Distances}{

* `calculate_distance`: Compute pairwise distances between two matrices
}

