% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_matrix.R
\name{expand_matrix}
\alias{expand_matrix}
\title{Expand a matrix with given rownames and colnames}
\usage{
expand_matrix(mat, rownames = NULL, colnames = NULL, fill = 0)
}
\arguments{
\item{mat}{The matrix to expand}

\item{rownames}{The desired rownames}

\item{colnames}{The desired colnames}

\item{fill}{With what to fill missing data}
}
\description{
Expand a matrix with given rownames and colnames
}
\examples{
x <- matrix(runif(12), ncol = 4, dimnames = list(c("a", "c", "d"), c("D", "F", "H", "I")))
expand_matrix(x, letters[1:5], LETTERS[1:10], fill = 0)
}
