% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_with.R
\name{extend_with}
\alias{extend_with}
\title{Extend an object}
\usage{
extend_with(object, .class_name, ...)
}
\arguments{
\item{object}{A list}

\item{.class_name}{A class name to add}

\item{...}{Extra information in the list}
}
\description{
Extend an object
}
