% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifna.R
\name{\%ifna\%}
\alias{\%ifna\%}
\title{Replace missing values}
\usage{
x \%ifna\% y
}
\arguments{
\item{x}{Original values, possibly containing \code{NA} values.}

\item{y}{Replacement values, length must be 1 or same as \code{x}.}
}
\description{
This function is similar to \link[rlang:\%|\%]{rlang::\%|\%} but expects the replacement values
to be of equal length to the initial vector.
}
