% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{has_names}
\alias{has_names}
\alias{\%has_names\%}
\title{Has names}
\usage{
has_names(x, which)

x \%has_names\% which
}
\arguments{
\item{x}{object to test}

\item{which}{name}
}
\description{
Has names
}
