% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_boilerplate.R
\name{inherit_default_params}
\alias{inherit_default_params}
\title{Allow Wouter to write less boilerplate}
\usage{
inherit_default_params(super_functions, fun)
}
\arguments{
\item{super_functions}{A list of super functions of which \code{fun} needs to inherit the default parameters}

\item{fun}{The function whose default parameters need to be overridden}
}
\value{
Function \code{fun}, but with the default parameters of the \code{super_functions}
}
\description{
A function to stop Wouter from crying.
}
\examples{

fun1 <- function(a = 10, b = 7) {
  runif(a, -b, b)
}

fun2 <- function(c = 9) {
  2^c
}

fun3 <- inherit_default_params(
  super = list(fun1, fun2),
  fun = function(a, b, c) {
    params <- as.list(environment())[formalArgs(fun3)]
    z <- fun1(a, b)
    y <- fun2(c)
    list(params = params, z = z, y = y)
  }
)
}
