% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{install_packages}
\alias{install_packages}
\title{Installs the suggests of a particular package including information from the remotes}
\usage{
install_packages(dependencies, package = NULL)
}
\arguments{
\item{dependencies}{The names of the packages to be installed}

\item{package}{The package from which the remotes will be extracted}
}
\description{
Installs the suggests of a particular package including information from the remotes
}
