% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_bounded}
\alias{is_bounded}
\title{Check whether a value within a certain interval}
\usage{
is_bounded(x, lower_bound = -Inf, lower_closed = FALSE,
  upper_bound = Inf, upper_closed = FALSE)
}
\arguments{
\item{x}{A value to be tested}

\item{lower_bound}{The lower bound}

\item{lower_closed}{Whether the lower bound is closed}

\item{upper_bound}{The upper bound}

\item{upper_closed}{Whether the upper bound is closed}
}
\description{
Check whether a value within a certain interval
}
