% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_single_numeric}
\alias{is_single_numeric}
\title{Check whether a value is a single numeric}
\usage{
is_single_numeric(x)
}
\arguments{
\item{x}{A value to be tested}
}
\description{
Check whether a value is a single numeric
}
