% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5.R
\name{is_sparse}
\alias{is_sparse}
\title{Check if a matrix is sparse}
\usage{
is_sparse(x)
}
\arguments{
\item{x}{A sparse or not sparse matrix}
}
\description{
Check if a matrix is sparse
}
\examples{
is_sparse(matrix(c(1:10)))
is_sparse(Matrix::Matrix(matrix(c(1:10))))
}
