% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_ti_data.R
\name{is_ti_data_wrapper}
\alias{is_ti_data_wrapper}
\title{Test whether something is a ti_data_wrapper or not}
\usage{
is_ti_data_wrapper(object)
}
\arguments{
\item{object}{The object to be tested.}
}
\description{
Test whether something is a ti_data_wrapper or not
}
