% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{manhattan_distance}
\alias{manhattan_distance}
\title{Manhattan distance between two matrices}
\usage{
manhattan_distance(x, y)
}
\arguments{
\item{x}{Matrix 1}

\item{y}{Matrix 2}
}
\description{
Implemented in Rcpp
}
