% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdf.R
\name{mapdf}
\alias{mapdf}
\alias{mapdf_lgl}
\alias{mapdf_chr}
\alias{mapdf_int}
\alias{mapdf_dbl}
\alias{mapdf_dfr}
\alias{mapdf_dfc}
\alias{walkdf}
\alias{mapdf_if}
\alias{mapdf_at}
\title{Apply a function to each row of a data frame}
\usage{
mapdf(.x, .f, ...)

mapdf_lgl(.x, .f, ...)

mapdf_chr(.x, .f, ...)

mapdf_int(.x, .f, ...)

mapdf_dbl(.x, .f, ...)

mapdf_dfr(.x, .f, ...)

mapdf_dfc(.x, .f, ...)

walkdf(.x, .f, ...)

mapdf_if(.x, .p, .f, ...)

mapdf_at(.x, .at, .f, ...)
}
\arguments{
\item{.x}{A \code{\link[base]{data.frame}}, \code{\link[tibble]{data_frame}} or \code{\link[tibble]{tibble}}.}

\item{.f}{A function in the form of \code{function(l, ...) {}},
where \code{l} will be one row from the data frame, converted as a list,
and \code{...} can be any other arguments passed to the \code{mapdf} function.}

\item{...}{Additional arguments passed on to \code{.f}.}

\item{.p}{A single predicate function, a formula describing such a
predicate function, or a logical vector of the same length as \code{.x}.
Alternatively, if the elements of \code{.x} are themselves lists of
objects, a string indicating the name of a logical element in the
inner lists. Only those elements where \code{.p} evaluates to
\code{TRUE} will be modified.}

\item{.at}{A character vector of names or a numeric vector of
positions. Only those elements corresponding to \code{.at} will be
modified.}
}
\description{
The mapdf functions transform their input by applying a function to each row of a data frame and returning a vector the same length as the input.
These functions work a lot like purrr's \code{\link[purrr]{map}} functions.
}
\details{
* \code{mapdf()}, \code{mapdf_if()} and \code{mapdf_at()} always return a list.

* \code{mapdf_lgl()}, \code{mapdf_int()}, \code{mapdf_dbl()} and \code{mapdf_chr()} return vectors of the corresponding type (or die trying).

* \code{mapdf_dfr()} and \code{mapdf_dfc()} return data frames created by row-binding and column-binding respectively. They require dplyr to be installed.

* \code{walkdf()} calls .f for its side-effect and returns the input .x.
}
\examples{
tib <- data.frame(x=1:3, y=3:5)
mapdf(tib, function(z) {z$x + z$y})
}
