% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{normalise_filter_counts}
\alias{normalise_filter_counts}
\title{State-of-the-art preprocessing and normalisation from https://f1000research.com/articles/5-2122/v2 and https://www.bioconductor.org/help/workflows/simpleSingleCell/}
\usage{
normalise_filter_counts(counts, has_spike = any(grepl("^ERCC",
  colnames(counts))), verbose = FALSE, nmads = 3,
  expressed_in_n_cells = 0.05, filter_hvg = TRUE, hvg_fdr = 0.05,
  hvg_bio = 0.5)
}
\arguments{
\item{counts}{The counts matrix, with genes in columns}

\item{has_spike}{Does this contain spike-ins, for which the gene names are preseded by ERCC}

\item{verbose}{Whether to add plots}

\item{nmads}{Number of median deviations for filtering outlier cells}

\item{expressed_in_n_cells}{Percentage of minimal number of cells a gene has to be expressed}

\item{filter_hvg}{Whether to filter out highly variable genes}

\item{hvg_fdr}{FDR gene filtering cutoff}

\item{hvg_bio}{Biological gene filtering cutoff}
}
\description{
State-of-the-art preprocessing and normalisation from https://f1000research.com/articles/5-2122/v2 and https://www.bioconductor.org/help/workflows/simpleSingleCell/
}
