% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynUtils-methods.R
\name{print.dynutillist}
\alias{print.dynutillist}
\title{Print Method for Object of Class \code{dynutillist}}
\usage{
\method{print}{dynutillist}(x, ...)
}
\arguments{
\item{x}{an object of class \code{dynutillist}.}

\item{...}{further arguments.}
}
\description{
Print Method for Object of Class \code{dynutillist}
}
\examples{
# prepare parameters
set.seed(42)
## number of individuals
n <- 5
## time points
time <- 5
## dynamic structure
p <- 3
mu0 <- rep(x = 0, times = p)
sigma0 <- 0.001 * diag(p)
sigma0_l <- t(chol(sigma0))
alpha <- rep(x = 0, times = p)
beta <- 0.50 * diag(p)
psi <- 0.001 * diag(p)
psi_l <- t(chol(psi))

library(simStateSpace)
ssm <- SimSSMVARFixed(
  n = n,
  time = time,
  mu0 = mu0,
  sigma0_l = sigma0_l,
  alpha = alpha,
  beta = beta,
  psi_l = psi_l,
  type = 0
)
data <- as.data.frame(ssm)
out <- SubsetByID(
  data = data,
  id = "id",
  time = "time",
  observed = paste0("y", 1:p)
)
print(out)

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
