% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_time_string.R
\name{random_time_string}
\alias{random_time_string}
\title{Generate random string}
\usage{
random_time_string(name = NULL)
}
\arguments{
\item{name}{Optional string to be added in the random_time_string}
}
\description{
Generate a random string with first the current time, together with a random number
}
\examples{
random_time_string("test")
}
