% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5.R
\name{read_h5}
\alias{read_h5}
\alias{write_h5}
\title{Read/write R objects to a H5 file.}
\usage{
read_h5(path)

write_h5(x, file)
}
\arguments{
\item{path}{Path to read from/write to.}

\item{x}{R object to write.}
}
\description{
Read/write R objects to a H5 file.
}
