% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_until_exit.R
\name{run_until_exit}
\alias{run_until_exit}
\title{Run until error
Run a system command until the R parent gets an exit signal, or until the command completes}
\usage{
run_until_exit(commands)
}
\arguments{
\item{commands}{Character scalar, the command to run.}
}
\description{
Run until error
Run a system command until the R parent gets an exit signal, or until the command completes
}
