% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_tempdir.R
\name{safe_tempdir}
\alias{safe_tempdir}
\title{Get temporary directory}
\usage{
safe_tempdir(subfolder)
}
\arguments{
\item{subfolder}{Name of a subfolder to be created}
}
\description{
Get temporary directory
}
