% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{scale_uniformily}
\alias{scale_uniformily}
\title{Scale data on a given center and between a given range}
\usage{
scale_uniformily(x, center =  0, max_range = 1)
}
\arguments{
\item{x}{A numeric matrix or data frame.}

\item{center}{The new center point of the data.}

\item{max_range}{The maximum range of each column.}
}
\value{
The centered and scaled matrix. The numeric centering and scalings used are returned as attributes.
}
\description{
\code{scale_uniformily} uniformily scales a given matrix such
that the returned space is centered on \code{center}, and each column was
scaled equally such that the range of each column is at most \code{max_range}.
}
