% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_network.R
\name{simplify_network}
\alias{simplify_network}
\title{Simplify a network, so that consecutive linear edges are merged}
\usage{
simplify_network(net)
}
\arguments{
\item{net}{A network in data.frame format containing at least the columns from, to and length}
}
\description{
Simplify a network, so that consecutive linear edges are merged
}
\examples{
net <- dplyr::tibble(from=1:2, to=2:3, length=1, directed=TRUE)
simplify_network(net)
}
