% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_network.R
\name{simplify_sample_graph}
\alias{simplify_sample_graph}
\title{Simplify a graph of samples}
\usage{
simplify_sample_graph(edges, is_trajectory, is_directed)
}
\arguments{
\item{edges}{The edges between samples. Should contain columns "from", "to", "length", and "directed"}

\item{is_trajectory}{A \strong{named} vector containing booleans fo whether or not a sample is to be kept.}

\item{is_directed}{Whether or not the graph is directed.}
}
\description{
Simplify a graph of samples
}
