% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_glue.R
\name{sticky_cat}
\alias{sticky_cat}
\title{Format, interpolate and print a string}
\usage{
sticky_cat(...)
}
\arguments{
\item{...}{\[`expressions`]\cr Expressions string(s) to format, multiple inputs are concatenated together before formatting.}
}
\description{
Uses \code{\link{pritt}} and \code{\link[base]{cat}}
}
