% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development.R
\name{switch_devel}
\alias{switch_devel}
\alias{switch_master}
\alias{switch_cran}
\title{Switching of development stage within the dynverse}
\usage{
switch_devel(file = "DESCRIPTION", desc = desc::desc(file = file))

switch_master(file = "DESCRIPTION", desc = desc::desc(file = file))

switch_cran(file = "DESCRIPTION", desc = desc::desc(file = file))
}
\arguments{
\item{file}{The description file, defaults to DESCRIPTION}

\item{desc}{The read in description using the desc package}
}
\description{
Switching of development stage within the dynverse
}
