% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5.R
\name{test_h5_installation}
\alias{test_h5_installation}
\alias{get_h5_test_data}
\title{Tests whether hdf5 is correctly installed and can load/write data}
\usage{
test_h5_installation(detailed = FALSE)

get_h5_test_data()
}
\arguments{
\item{detailed}{Whether top do a detailed check}
}
\description{
Tests whether hdf5 is correctly installed and can load/write data
}
