% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timings_list.R
\name{add_timing_checkpoint}
\alias{add_timing_checkpoint}
\alias{attach_timings_attribute}
\alias{get_timings_attribute}
\title{Helper functions for generating timings in dynmethods}
\usage{
add_timing_checkpoint(timings_list, name)

attach_timings_attribute(object, timings_list)

get_timings_attribute(object)
}
\arguments{
\item{timings_list}{The timings list of previous checkpoints}

\item{name}{The name of the timings checkpoint}

\item{object}{An object to attach the timings_list to, as a parameter}
}
\description{
Helper functions for generating timings in dynmethods
}
