% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{tmap}
\alias{tmap}
\title{Loop over each row in a tibble}
\usage{
tmap(.x, .f, ...)
}
\arguments{
\item{.x}{A list or atomic vector.}

\item{.f}{The function to apply to each row}

\item{...}{Additional arguments passed on to \code{.f}.}
}
\description{
Loop over each row in a tibble
}
\examples{
tib <- data.frame(x=1:3, y=3:5)
tmap(tib, function(z) {z$x + z$y})
}
