% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_list.R
\name{update_list}
\alias{update_list}
\title{Update a list based on another list (overwrite existing items, adds not existing items).
Does not work recursively.}
\usage{
update_list(x, y)
}
\arguments{
\item{x}{List to be updated}

\item{y}{List to update `x` with}
}
\value{
Update list
}
\description{
Update a list based on another list (overwrite existing items, adds not existing items).
Does not work recursively.
}
