% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.R
\name{update_news}
\alias{update_news}
\title{Update the news based on the md file}
\usage{
update_news(package = detect_package_name(), write = TRUE)
}
\arguments{
\item{package}{The package name}

\item{write}{Whether to overwrite news}
}
\description{
Update the news based on the md file
}
