% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_boilerplate.R
\name{updateList}
\alias{updateList}
\title{Update a list based on another list (overwrite existing items, adds not existing items).
Does not work recursively compared to modifyList}
\usage{
updateList(x, y)
}
\arguments{
\item{x}{List to be updated}

\item{y}{List to update `x` with}
}
\value{
Update list
}
\description{
Update a list based on another list (overwrite existing items, adds not existing items).
Does not work recursively compared to modifyList
}
