% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_ti_data.R
\name{wrap_linear_ti_prediction}
\alias{wrap_linear_ti_prediction}
\title{Wrap a linear TI prediction}
\usage{
wrap_linear_ti_prediction(id, cell_ids, pseudotimes, ...)
}
\arguments{
\item{id}{a unique identifier for the task}

\item{cell_ids}{the ids of the cells in the trajectory and counts}

\item{pseudotimes}{The pseudotimes of the \code{cell_ids}}

\item{...}{extra information to be stored in the wrapper}
}
\description{
This function will generate the milestone_network and progressions.
}
