% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_packages.R
\name{check_packages}
\alias{check_packages}
\title{Check which packages are installed}
\usage{
check_packages(...)
}
\arguments{
\item{...}{A set of package names}
}
\description{
Check which packages are installed
}
\examples{
check_packages("SCORPIUS", "dynutils")
check_packages(c("princurve", "mlr", "tidyverse"))
}
