% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynUtils-package.R
\docType{package}
\name{dynUtils-package}
\alias{dynUtils}
\alias{dynUtils-package}
\title{dynUtils: Dynamic Modeling Utilities}
\description{
Utility functions for data preparation when fitting dynamic models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeksterslab/dynUtils}
  \item \url{https://jeksterslab.github.io/dynUtils/}
  \item Report bugs at \url{https://github.com/jeksterslab/dynUtils/issues}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{r.jeksterslab@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]

}
\keyword{internal}
