% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{dynutils}
\alias{dynutils}
\alias{dynutils-package}
\title{Common functionality for the dynverse packages}
\description{
Provides common functionality for the dynverse packages.
dynverse is created to support the development, execution, and benchmarking of trajectory inference methods.
For more information, check out \url{https://github.com/dynverse/dynverse}.
}
\section{Manipulation of lists}{

\itemize{
  \item{\code{\link{add_class}}: Add a class to an object}
  \item{\code{\link{extend_with}}: Extend list with more data}
}
}

\section{Distances}{

\itemize{
  \item{\code{\link{calculate_distance}}: Compute pairwise distances between two matrices}
}
}

\section{Manipulation of matrices}{

\itemize{
  \item{\code{\link{expand_matrix}}: Add rows and columns to a matrix}
}
}

\section{Manipulation of functions}{

\itemize{
  \item{\code{\link{inherit_default_params}}: Have one function inherit the default parameters from other functions}
}
}

\section{Manipulation of packages}{

\itemize{
  \item{\code{\link{check_packages}}: Easily checking whether certain packages are installed}
  \item{\code{\link{install_packages}}: Install packages taking into account the remotes of another}
}
}

