% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeout.R
\name{eval_with_timeout}
\alias{eval_with_timeout}
\title{Wait for a handle to finish or kill it}
\usage{
eval_with_timeout(expr, envir = parent.frame(), timeout,
  on_timeout = c("error", "warning", "silent"))
}
\arguments{
\item{expr}{An R \code{\link[base]{expression}} to be evaluated.}

\item{envir}{The environment in which the expression should be evaluated.}

\item{timeout}{How long to wait before killing it.}

\item{on_timeout}{A character specifying what action to take if a timeout event occurs. Must be one of: error, warning, silent.}
}
\description{
Wait for a handle to finish or kill it
}
\examples{
\dontrun{
eval_with_timeout(
  expr = {
    Sys.sleep(100) # really long function
    data.frame(result = "finished", time = sleeper_time)
  },
  timeout = 10,
)
}
}
