% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_information.R
\name{generate_prior_information}
\alias{generate_prior_information}
\title{Extract the prior information from the milestone network}
\usage{
generate_prior_information(milestone_ids, milestone_network, progressions,
  milestone_percentages, counts, feature_info, cell_info)
}
\arguments{
\item{milestone_ids}{The milestone ids}

\item{milestone_network}{The milestone network}

\item{progressions}{The progressions}

\item{milestone_percentages}{The milestone percentages}

\item{counts}{The counts matrix}

\item{feature_info}{The feature info}

\item{cell_info}{The cell info}
}
\description{
For example, what are the start cells, the end cells, to which milestone does each cell belong to.
}
