% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{install_packages}
\alias{install_packages}
\title{Installs the suggests of a particular package including information from the remotes}
\usage{
install_packages(..., package = NULL, prompt = FALSE)
}
\arguments{
\item{...}{The names of the packages to be installed}

\item{package}{The package from which the remotes will be extracted. If used, this package does need to be installed.}

\item{prompt}{Whether to ask the user first for installation}
}
\description{
Installs the suggests of a particular package including information from the remotes
}
\examples{
\dontrun{
install_packages("SCORPIUS", package = "dynmethods", prompt = TRUE)
}
}
