% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{install_packages}
\alias{install_packages}
\title{Install packages, but first ask if interactive}
\usage{
install_packages(..., is_interactive = interactive())
}
\arguments{
\item{...}{The names of the packages to be installed}

\item{is_interactive}{Whether running interactivly, which will prompt the user before installation}
}
\description{
Install packages, but first ask if interactive
}
\examples{
\dontrun{
install_packages("SCORPIUS")
}
}
