% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{install_packages}
\alias{install_packages}
\title{Check package availability}
\usage{
install_packages(..., try_install = interactive())
}
\arguments{
\item{...}{The names of the packages to be checked}

\item{try_install}{Whether running interactivly, which will prompt the user before installation}
}
\description{
If the session is interactive, prompt the user whether to install the packages.
}
\examples{
\dontrun{
install_packages("SCORPIUS")
}
}
