% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_matrix.R
\name{is_sparse}
\alias{is_sparse}
\title{Check if an object is a sparse matrix}
\usage{
is_sparse(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Check if an object is a sparse matrix
}
\examples{
is_sparse(matrix(1:10)) # FALSE
is_sparse(Matrix::rsparsematrix(100, 200, .01)) # TRUE
}
