% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize_filter_counts}
\alias{normalize_filter_counts}
\title{State-of-the-art preprocessing and normalization from https://f1000research.com/articles/5-2122/v2}
\usage{
normalize_filter_counts(counts, has_spike = any(grepl("^ERCC",
  colnames(counts))), verbose = FALSE, nmads = 3,
  expressed_in_n_cells = 0.05, filter_hvg = TRUE, hvg_fdr = 0.05,
  hvg_bio = 0.5)
}
\arguments{
\item{counts}{The counts matrix, with genes in columns}

\item{has_spike}{Does this contain spike-ins, for which the gene names are preseded by ERCC}

\item{verbose}{Whether to add plots}

\item{nmads}{Number of median deviations for filtering outlier cells}

\item{filter_hvg}{Whether to filter out highly variable genes}
}
\description{
State-of-the-art preprocessing and normalization from https://f1000research.com/articles/5-2122/v2
}
