% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/override_setseed.R
\name{override_setseed}
\alias{override_setseed}
\title{Disable setting seeds}
\usage{
override_setseed(value)
}
\arguments{
\item{value}{an R object.}
}
\description{
This function was adapted from \code{assignInNamespace}, but
modified to only allow overwriting \code{set.seed}.
}
