% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processx_helpers.R
\name{print_processx}
\alias{print_processx}
\title{Processx callback function for printing stdout or stderr}
\usage{
print_processx(x, proc)
}
\arguments{
\item{x}{The stdout/stderr}

\item{proc}{The processx process}
}
\description{
Processx callback function for printing stdout or stderr
}
