% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pritt.R
\name{pritt}
\alias{pritt}
\title{Format and interpolate a string}
\usage{
pritt(...)
}
\arguments{
\item{...}{Parameters passed to \code{\link[glue]{glue}}.}
}
\description{
Uses \code{\link[glue]{glue}}, but removes the class from the output.
}
\seealso{
\code{\link[glue]{glue}}
}
