% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5.R
\name{read_h5}
\alias{read_h5}
\alias{read_h5_}
\alias{write_h5}
\alias{write_h5_}
\title{Read/write R objects to a H5 file.}
\usage{
read_h5(path)

read_h5_(file_h5)

write_h5(x, path)

write_h5_(x, file_h5, path)
}
\arguments{
\item{path}{Path to read from/write to.}

\item{file_h5}{A H5 file to read from/write to.}

\item{x}{R object to write.}
}
\description{
Read/write R objects to a H5 file.
}
