% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.R
\name{recent_news}
\alias{recent_news}
\title{Print the most recent news}
\usage{
recent_news(path = NULL, package = detect_package_name(path = path), n = 2)
}
\arguments{
\item{path}{The path of the description in which the package resides}

\item{package}{The package name}

\item{n}{Number of recent news to print}
}
\description{
Print the most recent news
}
