% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_tempdir.R
\name{safe_tempdir}
\alias{safe_tempdir}
\title{Create an empty temporary directory and return its path}
\usage{
safe_tempdir(subfolder)
}
\arguments{
\item{subfolder}{Name of a subfolder to be created}
}
\description{
Create an empty temporary directory and return its path
}
\examples{
\dontrun{
safe_tempdir("samson")
# "/tmp/Rtmp8xCGJe/file339a13bec763/samson"
}
}
