% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_add_cell_graph_to_wrapper.R
\name{simplify_sample_graph}
\alias{simplify_sample_graph}
\title{Add a linear trajectory to a data wrapper}
\usage{
simplify_sample_graph(data_wrapper, cell_graph, to_keep, is_directed, ...)
}
\arguments{
\item{data_wrapper}{A data wrapper to extend upon.}

\item{cell_graph}{The edges between cells. Format: Data frame(from = character, to = character, length = numeric)}

\item{to_keep}{A named vector containing booleans containing
whether or not a sample is part of the trajectory that is to be kept.}

\item{is_directed}{Whether or not the graph is directed.}

\item{...}{extra information to be stored in the wrapper.}
}
\description{
This function will generate the milestone_network and progressions.
}
